+++ src/ai_identicon/__init__.py
@@ -36,7 +36,7 @@ from .model import AvatarModel, AvatarState
 from .portrait import color_svg, export_svg, line_art_svg
 from .controller import AvatarController
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
     "__version__",